+++ src/Twilio/Rest/Accounts/V1.php
@@ -22,6 +22,7 @@ use Twilio\Rest\Accounts\V1\AuthTokenPromotionList;
 use Twilio\Rest\Accounts\V1\BulkConsentsList;
 use Twilio\Rest\Accounts\V1\BulkContactsList;
 use Twilio\Rest\Accounts\V1\CredentialList;
+use Twilio\Rest\Accounts\V1\MessagingGeopermissionsList;
 use Twilio\Rest\Accounts\V1\SafelistList;
 use Twilio\Rest\Accounts\V1\SecondaryAuthTokenList;
 use Twilio\Version;
@@ -31,6 +32,7 @@ use Twilio\Version;
  * @property BulkConsentsList $bulkConsents
  * @property BulkContactsList $bulkContacts
  * @property CredentialList $credentials
+ * @property MessagingGeopermissionsList $messagingGeopermissions
  * @property SafelistList $safelist
  * @property SecondaryAuthTokenList $secondaryAuthToken
  */
@@ -40,6 +42,7 @@ class V1 extends Version
     protected $_bulkConsents;
     protected $_bulkContacts;
     protected $_credentials;
+    protected $_messagingGeopermissions;
     protected $_safelist;
     protected $_secondaryAuthToken;
 
@@ -84,6 +87,14 @@ class V1 extends Version
             $this->_credentials = new CredentialList($this);
         }
         return $this->_credentials;
+    }
+
+    protected function getMessagingGeopermissions(): MessagingGeopermissionsList
+    {
+        if (!$this->_messagingGeopermissions) {
+            $this->_messagingGeopermissions = new MessagingGeopermissionsList($this);
+        }
+        return $this->_messagingGeopermissions;
     }
 
     protected function getSafelist(): SafelistList