+++ src/Twilio/Rest/Api/V2010/Account/Conference/ParticipantOptions.php
@@ -67,7 +67,7 @@ abstract class ParticipantOptions
      * @param string $amdStatusCallbackMethod The HTTP method we should use when calling the `amd_status_callback` URL. Can be: `GET` or `POST` and the default is `POST`.
      * @param string $trim Whether to trim any leading and trailing silence from the participant recording. Can be: `trim-silence` or `do-not-trim` and the default is `trim-silence`.
      * @param string $callToken A token string needed to invoke a forwarded call. A call_token is generated when an incoming call is received on a Twilio number. Pass an incoming call's call_token value to a forwarded call via the call_token parameter when creating a new call. A forwarded call should bear the same CallerID of the original incoming call.
-     * @param string $callerDisplayName The name that appears to the called party for this call. Must be between 2 and 255 characters.
+     * @param string $callerDisplayName The name that populates the display name in the From header. Must be between 2 and 255 characters. Only applicable for calls to sip address.
      * @return CreateParticipantOptions Options builder
      */
     public static function create(
@@ -295,7 +295,7 @@ class CreateParticipantOptions extends Options
      * @param string $amdStatusCallbackMethod The HTTP method we should use when calling the `amd_status_callback` URL. Can be: `GET` or `POST` and the default is `POST`.
      * @param string $trim Whether to trim any leading and trailing silence from the participant recording. Can be: `trim-silence` or `do-not-trim` and the default is `trim-silence`.
      * @param string $callToken A token string needed to invoke a forwarded call. A call_token is generated when an incoming call is received on a Twilio number. Pass an incoming call's call_token value to a forwarded call via the call_token parameter when creating a new call. A forwarded call should bear the same CallerID of the original incoming call.
-     * @param string $callerDisplayName The name that appears to the called party for this call. Must be between 2 and 255 characters.
+     * @param string $callerDisplayName The name that populates the display name in the From header. Must be between 2 and 255 characters. Only applicable for calls to sip address.
      */
     public function __construct(
         
@@ -950,9 +950,9 @@ class CreateParticipantOptions extends Options
     }
 
     /**
-     * The name that appears to the called party for this call. Must be between 2 and 255 characters.
+     * The name that populates the display name in the From header. Must be between 2 and 255 characters. Only applicable for calls to sip address.
      *
-     * @param string $callerDisplayName The name that appears to the called party for this call. Must be between 2 and 255 characters.
+     * @param string $callerDisplayName The name that populates the display name in the From header. Must be between 2 and 255 characters. Only applicable for calls to sip address.
      * @return $this Fluent Builder
      */
     public function setCallerDisplayName(string $callerDisplayName): self