+++ src/Twilio/Rest/Numbers/V1/PortingPortInInstance.php
@@ -41,6 +41,7 @@ use Twilio\Deserialize;
  * @property string|null $autoCancelApprovalNumbers
  * @property string[]|null $documents
  * @property \DateTime|null $dateCreated
+ * @property int $supportTicketId
  */
 class PortingPortInInstance extends InstanceResource
 {
@@ -73,6 +74,7 @@ class PortingPortInInstance extends InstanceResource
             'autoCancelApprovalNumbers' => Values::array_get($payload, 'auto_cancel_approval_numbers'),
             'documents' => Values::array_get($payload, 'documents'),
             'dateCreated' => Deserialize::dateTime(Values::array_get($payload, 'date_created')),
+            'supportTicketId' => Values::array_get($payload, 'support_ticket_id'),
         ];
 
         $this->solution = ['portInRequestSid' => $portInRequestSid ?: $this->properties['portInRequestSid'], ];